+++ src/Revenue/Migrations/AddPastDonationsToRevenueTable.php
@@ -3,6 +3,7 @@ namespace Give\Revenue\Migrations;
 
 use Give\Framework\Migrations\Contracts\Migration;
 use Give\Revenue\Repositories\Revenue;
+use Give\ValueObjects\Money;
 use Give_Updates;
 use InvalidArgumentException;
 use WP_Query;
@@ -63,10 +64,14 @@ class AddPastDonationsToRevenueTable extends Migration {
 					continue;
 				}
 
+				if ( ! ( $amount = give()->payment_meta->get_meta( $post->ID, '_give_cs_base_amount', true ) ) ) {
+					$amount = give_donation_amount( $post->ID );
+				}
+
 				$revenueData = [
 					'donation_id' => $post->ID,
 					'form_id'     => give_get_payment_form_id( $post->ID ),
-					'amount'      => give_donation_amount( $post->ID ) * 100,
+					'amount'      => Money::of( $amount, give_get_option( 'currency' ) )->getMinorAmount(),
 				];
 
 				try {